server/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
bcrypt
fastapi-sessions
pydantic
